+++ docs/changelog.md
@@ -17,6 +17,10 @@ Types of changes
 
 ## Unreleased
 
+Added:
+
+- The rendering of explorer tiles can now be fully customized. For each kind of special tile (e.g. visited, missing, new in this activity), you can set a fill color, border color, border style and stripe color. ([GH-366](https://github.com/martin-ueding/geo-activity-playground/issues/366))
+
 
 ## Version 1.46.0 — 2026-08-03